training cadence contorl through dt

+++ samples/training/scripts/run_submit.sh
@@ -32,9 +32,10 @@ apptainer exec --nv \
 		-B $snap_src:$sample_src/snapshots \
 		torchfort_latest_withdf5.sif bash /opt/scripts/run.sh
 
-cp stats_0.pt $data_src/training
+if [ -f "stats_0.pt" ]; then
+	cp stats_0.pt $data_src/training
 
-module load python-data
-module load pytorch/2.4
+	module load python-data
+	module load pytorch/2.4
 
-srun python3 $data_src/training/ptTobin.py
+	srun python3 $data_src/training/ptTobin.py
